5_Symbols/rag/requirements.txt
faiss-cpu==1.7.4
sentence-transformers==2.2.2
markdown-it-py==3.0.0
numpy==1.24.3
argparse